Fixed calories calculation, added water category and added progress bar

+++ lib/models/pdf_items_model.dart
@@ -7,7 +7,7 @@ class ItemModel {
   TextEditingController? itemQty;
   bool? isEnableEditing;
   String? itemQtyNum;
-  int? totalCal;
+  double? totalCal;
 
   ItemModel({
     this.itemName,
